keycloak_fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-property==0.2.2
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
cryptography==42.0.8
deprecation==2.1.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.12
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
locust==2.25.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
orjson==3.10.3
packaging==24.1
psutil==5.9.8
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.3
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==4.2.0
python-multipart==0.0.9
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
roundrobin==0.0.4
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlmodel==0.0.19
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
